feat(message-content): enhance output_text handling and add tests for message part extraction
Improved the handling of `output_text` parts in conversation messages by ensuring they are preserved during extraction and normalization. Updated the `normalizeConversationContentPart` function to include `normalizePersistedOutputTextPart`. Added tests to validate the correct mapping of `output_text` parts in various scenarios, ensuring they are accurately represented in the UI messages.
