+++ apps/web/src/app/(app)/chat/utils/__tests__/message-utils.test.ts
@@ -1,75 +1,6 @@
 import { describe, expect, it } from "vitest";
 
-import {
-  convertItemsToMessages,
-  deduplicateMessagesById,
-} from "../message-utils";
-
-describe("convertItemsToMessages", () => {
-  it("maps API content arrays with reasoning then assistant text", () => {
-    const messages = convertItemsToMessages([
-      {
-        id: "a1",
-        role: "assistant",
-        createdAt: 1700000000,
-        content: [
-          { type: "reasoning", text: "Think" },
-          { type: "output_text", text: "Hi" },
-        ],
-      },
-    ]);
-    expect(messages[0]?.parts).toEqual([
-      { type: "reasoning", text: "Think" },
-      { type: "text", text: "Hi" },
-    ]);
-  });
-
-  it("maps provider-specific reasoning types to reasoning parts, not visible text", () => {
-    const messages = convertItemsToMessages([
-      {
-        id: "a2",
-        role: "assistant",
-        createdAt: 1700000000,
-        content: [
-          { type: "redacted_reasoning", text: "[hidden]" },
-          { type: "output_text", text: "Hi" },
-        ],
-      },
-    ]);
-    expect(messages[0]?.parts).toEqual([
-      { type: "reasoning", text: "[hidden]" },
-      { type: "text", text: "Hi" },
-    ]);
-  });
-
-  it("maps file parts from API content arrays", () => {
-    const messages = convertItemsToMessages([
-      {
-        id: "u1",
-        role: "user",
-        createdAt: 1700000001,
-        content: [
-          { type: "text", text: "See attached" },
-          {
-            type: "file",
-            url: "https://example.com/blob.png",
-            mediaType: "image/png",
-            filename: "blob.png",
-          },
-        ],
-      },
-    ]);
-    expect(messages[0]?.parts).toEqual([
-      { type: "text", text: "See attached" },
-      {
-        type: "file",
-        url: "https://example.com/blob.png",
-        mediaType: "image/png",
-        filename: "blob.png",
-      },
-    ]);
-  });
-});
+import { deduplicateMessagesById } from "../message-utils";
 
 describe("deduplicateMessagesById", () => {
   it("keeps every message when ids are missing (no empty-key collapse)", () => {
